pyproject.toml
[tool.poetry]
name = "jobby"
version = "0.1.0"
description = ""
authors = ["Ryan Wiesenberg <rmwiesenberg@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.3"
requests = "^2.32.3"
pyyaml = "^6.0.2"
tqdm = "^4.67.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
